refactor: move AutoOrderSelector to utils and remove model_selection folder
- Move AutoOrderSelector from model_selection/best_lag.py to utils/auto_order_selector.py
- Remove empty model_selection folder entirely
- Update all imports to use new location
- Rename test_best_lag.py to test_auto_order_selector_legacy.py for clarity
- Export AutoOrderSelector from utils/__init__.py
- Fix all patch statements in tests to use new import path

This simplifies the project structure by removing a folder that contained
only one file with a confusing name (best_lag.py).
