--- src/tsbootstrap/model_selection/__init__.py
@@ -1,5 +1,0 @@
-"""Model selection utilities for tsbootstrap."""
-
-from .best_lag import AutoOrderSelector
-
-__all__ = ["AutoOrderSelector"]
