+++ src/tsbootstrap/utils/__init__.py
@@ -1,5 +1,6 @@
 """Utilities for tsbootstrap package."""
 
+from tsbootstrap.utils.auto_order_selector import AutoOrderSelector
 from tsbootstrap.utils.estimator_checks import check_estimator
 
-__all__ = ["check_estimator"]
+__all__ = ["AutoOrderSelector", "check_estimator"]
