+++ src/tsbootstrap/utils/auto_order_selector.py
